Refactor agents with absolute imports and standalone services
