+++ src/agents/reporting/reporting_agent.py
@@ -1,0 +1,9 @@
+from src.templates import BaseAgent
+
+
+class ReportingAgent(BaseAgent):
+    """Generates reports and dashboards."""
+
+    def report(self, data) -> dict:
+        """Return a simple summary report."""
+        return {"total_items": len(data)}
