+++ src/orchestrator/__init__.py
@@ -50,3 +50,6 @@ class Orchestrator:
             "tickets": tickets,
             "report": report,
         }
+
+
+__all__ = ["Orchestrator"]
