+++ src/orchestrator/agents/triage/__init__.py
@@ -1,15 +1,3 @@
-try:
-    from src.templates import BaseAgent
-except ModuleNotFoundError:  # allow running standalone
-    from templates import BaseAgent
+from src.orchestrator.agents.triage.triage_agent import TriageAgent
 
-
-class TriageAgent(BaseAgent):
-    """Assigns risk scores using Azure OpenAI."""
-
-    def triage(self, finding: dict):
-        """Return finding with a dummy risk score."""
-        score = len(str(finding)) % 5
-        result = dict(finding)
-        result["risk_score"] = score
-        return result
+__all__ = ["TriageAgent"]
